backend/requirements.txt
asgiref==3.8.1
Django==5.1.5
djangorestframework==3.15.2
psycopg2==2.9.10
sqlparse==0.5.3
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.14.0